fix(2fa): updated auth messgae, upated routes for 2fa, thrown errors properly
